Refactor for consistency
