--- src/Faker.NET/Airline/Data/Airlines.cs
@@ -1,21 +1,0 @@
-namespace Faker.NET.Airline.Data;
-
-public class Airlines
-{
-    public static readonly Airline[] MainlineAirlines = {
-        new Airline("Alaska Airlines","AS"),
-        new Airline("Allegiant Air","G4"),
-        new Airline("American Airlines","AA"),
-        new Airline("Avelo Airlines","XP"),
-        new Airline("Breeze Airways","MX"),
-        new Airline("Delta Air Lines","DL"),
-        new Airline("Eastern Airlines","2D"),
-        new Airline("Frontier Airlines","F9"),
-        new Airline("Hawaiian Airlines","HA"),
-        new Airline("JetBlue","B6"),
-        new Airline("Southwest Airlines","WN"),
-        new Airline("Spirit Airlines","NK"),
-        new Airline("Sun Country Airlines","SY"),
-        new Airline("United Airlines","UA"),
-    };
-}
