+++ src/Faker.NET/Common/Airline/AircraftType.cs
@@ -1,4 +1,4 @@
-namespace Faker.NET.Airline;
+namespace Faker.NET.Common.Airline;
 
 public enum AircraftType
 {
