+++ src/Faker.NET/Common/Airline/Airline.cs
@@ -1,4 +1,4 @@
-namespace Faker.NET.Airline;
+namespace Faker.NET.Common.Airline;
 
 public class Airline
 {
