+++ src/Faker.NET/Common/Airline/Airport.cs
@@ -1,4 +1,4 @@
-namespace Faker.NET.Airline;
+namespace Faker.NET.Common.Airline;
 
 public class Airport
 {
